# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Software migration management module",
        "title": "CM-Grasshopper REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/grasshopper",
    "paths": {
...
        "/velero/migration/execute": {
            "post": {
                "description": "Run source backup and target restore as a single asynchronous Velero migration job.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Execute Migration",
                "operationId": "velero-migration-execute",
                "parameters": [
                    {
                        "description": "Velero migration execute request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully started migration job.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobExecution"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to start migration job.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/migration/precheck": {
            "post": {
                "description": "Check source cluster, target cluster, and MinIO before executing Velero migration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Precheck Migration",
                "operationId": "velero-migration-precheck",
                "parameters": [
                    {
                        "description": "Velero migration precheck request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationPrecheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully completed precheck.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroPrecheckResponse"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to run migration precheck.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/source/backups": {
            "post": {
                "description": "Create Velero backup on source cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Create Backup",
                "operationId": "create-velero-backup",
                "parameters": [
                    {
                        "description": "Velero backup request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created backup.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create backup.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/source/backups/list": {
            "post": {
                "description": "List Velero backups on source cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "List Backups",
                "operationId": "list-velero-backups",
                "parameters": [
                    {
                        "description": "Source cluster access request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.MultiClusterEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully listed backups.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to list backups.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/source/backups/{name}": {
            "post": {
                "description": "Get Velero backup detail on source cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Get Backup",
                "operationId": "get-velero-backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup name.",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Source cluster access request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.MultiClusterEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully got backup detail.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get backup detail.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/source/backups/{name}/delete": {
            "post": {
                "description": "Delete Velero backup on source cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Delete Backup",
                "operationId": "delete-velero-backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup name.",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Source cluster access request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.MultiClusterEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted backup.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.SimpleMsg"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete backup.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/velero/source/backups/{name}/validate": {
            "post": {
                "description": "Validate Velero backup status on source cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] Velero migration APIs"
                ],
                "summary": "Validate Backup",
                "operationId": "validate-velero-backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup name.",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Source cluster access request.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.MultiClusterEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully validated backup.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to validate backup.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.SimpleMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationExecuteRequest": {
            "type": "object",
            "properties": {
                "migration": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationExecuteSpec"
                },
                "sourceCluster": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.ClusterAccess"
                },
                "storage": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.StorageAccess"
                },
                "targetCluster": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.ClusterAccess"
                }
            }
        },
...
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationPrecheckRequest": {
            "type": "object",
            "properties": {
                "precheck": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.VeleroMigrationPrecheckSpec"
                },
                "sourceCluster": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.ClusterAccess"
                },
                "storage": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.StorageAccess"
                },
                "targetCluster": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.ClusterAccess"
                }
            }
        },
...
    }
}